# docs/docs/specs/express-openapi.json
{
...
    "openapi": "3.0.3",
    "info": {
        "title": "InsureCRM API",
        "version": "1.0.0",
        "description": "REST API for InsureCRM \u2014 a CRM platform for insurance agencies. Manages authentication, client records, and policy tracking."
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
...
    "paths": {
...
        "/api/notifications": {
            "post": {
                "summary": "Create a notification for a client",
                "description": "Creates a new notification and associates it with a client. The notification is assigned a unique ID (NOTIF-XXXXXX) and starts in unread state. If priority is not specified, it defaults to \"medium\".\n",
                "tags": [
                    "Notifications"
                ],
                "operationId": "createNotification",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateNotificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Notification"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required fields \u2014 clientId, type, title, and message are all required"
                    }
                }
            },
            "get": {
                "summary": "List notifications with filters and pagination",
                "description": "Returns a paginated list of notifications, sorted by creation date (newest first). Supports filtering by client, notification type, read status, and priority level.\n",
                "tags": [
                    "Notifications"
                ],
                "operationId": "listNotifications",
                "parameters": [
                    {
                        "in": "query",
                        "name": "clientId",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Filter to notifications belonging to this client"
                    },
                    {
                        "in": "query",
                        "name": "type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "claim_update",
                                "policy_renewal",
                                "payment_due",
                                "document_request",
                                "general"
                            ]
                        },
                        "description": "Filter by notification category"
                    },
                    {
                        "in": "query",
                        "name": "unreadOnly",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false"
                            ]
                        },
                        "description": "When \"true\", return only unread notifications"
                    },
                    {
                        "in": "query",
                        "name": "priority",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high",
                                "urgent"
                            ]
                        },
                        "description": "Filter by priority level"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number for pagination"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        },
                        "description": "Number of notifications per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "notifications": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notification"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "description": "Total number of matching notifications"
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "limit": {
                                            "type": "integer"
                                        },
                                        "totalPages": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/{notificationId}": {
            "get": {
                "summary": "Get a single notification by ID",
                "description": "Retrieves the full details of a specific notification by its unique identifier.",
                "tags": [
                    "Notifications"
                ],
                "operationId": "getNotification",
                "parameters": [
                    {
                        "in": "path",
                        "name": "notificationId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Unique notification ID (e.g. NOTIF-000001)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Notification"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No notification found with the given ID"
                    }
                }
            },
            "delete": {
                "summary": "Delete a notification",
                "description": "Permanently removes a notification. This action cannot be undone.",
                "tags": [
                    "Notifications"
                ],
                "operationId": "deleteNotification",
                "parameters": [
                    {
                        "in": "path",
                        "name": "notificationId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Unique notification ID (e.g. NOTIF-000001)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification deleted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No notification found with the given ID"
                    }
                }
            }
        },
        "/api/notifications/{notificationId}/read": {
            "patch": {
                "summary": "Mark a notification as read",
                "description": "Sets the notification's read flag to true and records the current timestamp as readAt. Idempotent \u2014 marking an already-read notification has no adverse effect.\n",
                "tags": [
                    "Notifications"
                ],
                "operationId": "markNotificationRead",
                "parameters": [
                    {
                        "in": "path",
                        "name": "notificationId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Unique notification ID (e.g. NOTIF-000001)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Notification"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No notification found with the given ID"
                    }
                }
            }
        },
        "/api/notifications/client/{clientId}/read-all": {
            "patch": {
                "summary": "Mark all notifications for a client as read",
                "description": "Bulk-marks every unread notification belonging to the specified client as read. Returns the count of notifications that were updated. Already-read notifications are not affected.\n",
                "tags": [
                    "Notifications"
                ],
                "operationId": "markAllClientNotificationsRead",
                "parameters": [
                    {
                        "in": "path",
                        "name": "clientId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Client ID whose unread notifications should be marked as read (e.g. cli_001)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All unread notifications for the client marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marked 3 notifications as read"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "description": "Number of notifications that were marked as read"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Notification": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique notification identifier",
                        "example": "NOTIF-000001"
                    },
                    "clientId": {
                        "type": "string",
                        "description": "ID of the client this notification belongs to",
                        "example": "cli_001"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "claim_update",
                            "policy_renewal",
                            "payment_due",
                            "document_request",
                            "general"
                        ],
                        "description": "Category of the notification"
                    },
                    "title": {
                        "type": "string",
                        "description": "Short headline for the notification",
                        "example": "Policy Renewal Reminder"
                    },
                    "message": {
                        "type": "string",
                        "description": "Full notification body text",
                        "example": "Your auto policy SF-2024-00123 is due for renewal on 2025-01-01."
                    },
                    "read": {
                        "type": "boolean",
                        "description": "Whether the notification has been read"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "description": "Urgency level of the notification"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "ISO 8601 timestamp when the notification was created"
                    },
                    "readAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true,
                        "description": "ISO 8601 timestamp when the notification was read, or null if unread"
                    },
                    "metadata": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "description": "Optional key-value pairs for additional context (e.g. policyNumber, claimId)"
                    }
                }
            },
            "CreateNotificationRequest": {
                "type": "object",
                "required": [
                    "clientId",
                    "type",
                    "title",
                    "message"
                ],
                "properties": {
                    "clientId": {
                        "type": "string",
                        "description": "ID of the client to notify",
                        "example": "cli_001"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "claim_update",
                            "policy_renewal",
                            "payment_due",
                            "document_request",
                            "general"
                        ],
                        "description": "Category of the notification"
                    },
                    "title": {
                        "type": "string",
                        "description": "Short headline for the notification",
                        "example": "Policy Renewal Reminder"
                    },
                    "message": {
                        "type": "string",
                        "description": "Full notification body text",
                        "example": "Your auto policy is due for renewal next month."
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "description": "Urgency level (defaults to \"medium\" if omitted)"
                    },
                    "metadata": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "description": "Optional key-value pairs for additional context"
                    }
                }
            }
        }
    }
}